backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-storages==1.14.2
boto3==1.29.7
psycopg2-binary==2.9.9
tenacity==8.2.3
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-django==4.7.0
requests==2.31.0
Pillow==10.1.0
